Initial commit of cppcodec.
It's starting out with a README.md that contains a pretty
comprehensive documentation of the API and design choices,
plus a first crack at implementing Crockford base32 within
that framework.

The base32 implementation might be incorrect and will have to
be verified/tested, however it exercises all the infrastructure
to go from app to (templated) interface to codec. So adding more
codecs should be fairly straightforward without too much effort.