Refine base32 accepted characters, rename crockstr to crockford.
For naming, I decided to let place-based single number codecs
have the original octet-streaming name plus "_num" appended.
Because they're an edge case, they'll perform worse, and they
people shouldn't be encouraged to use them. A slightly
inconvenient suffix will do that and make the rest of the names
easier to read and come up with.

As for accepted characters, Crockford base32 was missing special
treatment for 'OoIiLl' and RFC 4648 base32 was not accepting
lower-case characters, although it should. Both are now fixed.

README.txt got a bit of a rewrite for codec variants so all of
the codecs are now listed and with hopefully better explanations.