Make the Hello World example from README.md actually work.
Since it passes a string literal (a.k.a. constexpr char array),
we have to make sure to support that for size() and char_data().

Along the way I picked up another piece of template magic too,
to make put() not fail in one instance. As a nice side effect,
this allows us to get rid of the function_traits file.
I'm still only beginning to understand how it could possibly work,
but it seems to be more robust than the previous one.