Improve decoding performance by always inlining Codec::index_of()
In the gaspardpetit/base64 test, perf showed a significant amount
of time spent in that function, which seemed off because that
function should not exist at all in a release build.

Making sure it's inline reduces decoding time for Gaspard's 256
workload by almost 30% (from approx. 2.38 to 1.69 on my machine).
It seems logical to assume that other codecs get a similar
performance boost this way, so I'm applying the change everywhere.

This puts performance characteristics around the same level of
adamvr/arduino-base64 or the Manuel Martinez one (but without
the latter's apparent setup overhead). Still not the best,
but now definitely in the "fair" spectrum for decoding.