Update Travis config to use Trusty and several C++11 standards
Since we're now on Trusty, there are more and cleaner options
for testing clang, so the test matrix is expanded by adding
several extra clang builds.

I tried to put in a good mix of C++11, C++14 and C++17 builds.
The former two ought to behave the same, whereas C++17 has a
change to std::string::data() which causes a different optimized
code path to be selected. Testing it with C++17 will verify the
build fix for #41 from my previous commit.

I also believe that passing the CMAKE_BUILD_TYPE didn't work for
selecting Debug/Release configurations, which is fixed now by
passing it as a command line argument instead of merely setting
an environment variable.