Add a benchmark to test runtime performance.
Numbers are comparable to the gaspardpetit/base64 benchmark,
in which the runtime duration of encoding or decoding runs is
averaged out over many iterations (1 million or up to 0.5s,
whichever one is shorter). Code is all new.

Run test/benchmark_cppcodec without arguments to get numbers
for encoding to strings, decoding to strings and decoding to
vector<uint8_t> for base64_rfc4648.

This should be enough to test the performance fix for Windows
from #35. We can add command line options later, such as
selecting the codec or the set of decoded buffer sizes.