Execute test/benchmark_cppcodec on Travis and AppVeyor builds.
We can't be sure if the resources for the builds are going to be
consistent, so performance numbers may not be comparable from one
build to another. However, they should get us a rough idea about
differences on an order of magnitude, and/or average out over
multiple builds. If things are going very wrong or very right,
the build bots will help provide indicators.