Fix #35: Improve MSVC performance by adding MSVC always-inline macro
GCC and Clang have long had this optimization, which reduces
function calls in the generated assembly for (template) code
that I wanted to make more modular but without introducing
actual physical indirections. For the Visual Studio compiler,
I hadn't previously been able to verify that it works so I
didn't hadn't incorporated this change until now.

For the current codebase, this makes a particularly large impact
on encoding performance. MSVC encoding runtime on my machine goes
down from 1.90 to 1.00 for buffer size 256, which is a 52%
reduction in time spent. Decoding on MSVC for the same buffer size
was 1.95 and is now 1.46, i.e. 25% less time than before.

It seems like MSVC (on Windows) is still slower than GCC or Clang
(on Linux) at least on my system, where my numbers for the same
benchmark were 0.85 for encoding and 0.80 for decoding according
to an earlier commit message. Travis's benchmark results are much
closer to the size-256 results for MSVC, which might be relevant
because my Linux numbers for large buffer sizes (size 32768) are
roughly similar to Travis's.

MSVC improvements for size 32768 are roughly proportional for
encoding (222 -> 98) but less pronounced for decoding (533 -> 498),
implying that the latter is still dragged down by something else.