Compile on MSVC with /std:c++17 even without /Zc:__cplusplus
MSVC up to at least VS 2017 would define __cplusplus as 199711L
even when a different standard is passed as command line option,
because it hasn't been fully compatible with C++11 (and later)
requirements so far. To rectify this, they introduced another
option that fixes the __cplusplus macro, which is how AppVeyor
and others were able to build the code at all.

By adding an extra check for the Microsoft-specific macro
_MSVC_LANG, it's possible to build even without passing the
extra option.

See also:
https://docs.microsoft.com/en-us/cpp/preprocessor/predefined-macros
https://blogs.msdn.microsoft.com/vcblog/2018/04/09/msvc-now-correctly-reports-__cplusplus/