Change the LICENSE file so GitHub will recognize and highlight its info.
This includes changing it from LICENSE.txt to LICENSE,
and adding a more conventional copyright header instead of the
current, more verbose one.

One-time contributors are still listed with full copyright
in the individual codec variant files. I chose not to include
those here for maintainability reasons, but if you're one of them
and you'd like to show up in here as well, let me know and it
shall be done.