Un-recommend the baseNN_default_* headers.
They introduce complexity, make autocompletion more cumbersome,
and shouldn't be included in other header files because they
pollute the global namespace. If people copy base64 code from
one file to the other, the meaning of the code can change based
on which default header is included, which is not great.

I thought they added convenience, but actually, including them
makes cppcodec worse (probably).

So the first thing I can do is to stop recommending these headers.
The second thing will be to deprecate them. This may or may not
happen, I have to contemplate if the work inflicted on users for
changing header files outweighs the benefit of not having too many
"wrong" templates in other GitHub projects.

Either way, no new default headers will be added to cppcodec.