Add CPPCODEC_ALWAYS_INLINE statements to raw_result_buffer.
This should make sure it doesn't suffer from unnecessary slowness.
The good news is that GCC and Clang release builds don't seem to
be noticeably affected by this change, they must have already
done the right thing (inlining) beforehand. Nonetheless, compilers
with less predictive skill may get a speed-up from this.

Also, remove raw_result_buffer::last() because it's not needed
and makes it look like it's necessary for data wrappers, which
it's not.